move jest to ava, ts-node to esbuild, remove unused deps
