+++ ava.config.js
@@ -1,0 +1,5 @@
+module.exports = {
+  require: ['esbuild-register'],
+  extensions: ['ts'],
+  files: ['./tests/**/*.test.ts'],
+}
