+++ tsconfig.json
@@ -17,7 +17,6 @@
     "noLib": false,
     "skipLibCheck": true,
     "jsx": "react",
-    "types": ["node", "jest"],
     "lib": ["dom", "es2015", "es2016", "es2017", "esnext.asynciterable"],
     "keyofStringsOnly": true,
